Fixed error: integer literal '4294967295' overflows when stored into 'Int'.

+++ modules/RoxieMobile.SwiftCommons/Core.Extensions/Module/Sources/Roxie/Roxie.NewTag.swift
@@ -1,6 +1,6 @@
 // ----------------------------------------------------------------------------
 //
-//  Roxie.CustomTag.swift
+//  Roxie.NewTag.swift
 //
 //  @author     Alexander Bragin <bragin-av@roxiemobile.com>
 //  @copyright  Copyright (c) 2017, Roxie Mobile Ltd. All rights reserved.
@@ -20,8 +20,8 @@ public extension Roxie
     /// Creates a new URN tag for any object.
     public static func newTag(for subject: Any) -> String
     {
-        let nameTag = String(format:"%08lx", Roxie.typeName(of: subject).hashValue & 0xFFFFFFFF)
-        let timeTag = String(format:"%08lx", mach_absolute_time() & 0xFFFFFFFF)
+        let nameTag = String(format:"%08lx", UInt32(Roxie.typeName(of: subject).hashValue) & 0xFFFFFFFF)
+        let timeTag = String(format:"%08lx", UInt32(mach_absolute_time()) & 0xFFFFFFFF)
 
         // Done
         return "urn:tag:\(nameTag):\(timeTag)"
